#ORM_Refactoring
Sumit - Fixed defects below defects -
1. Add/Update/Delete Status/Response Code.
2. Rules Fetch.

+++ src/main/java/com/mcmcg/ico/bluefin/service/InternalResponseCodeService.java
@@ -367,6 +367,9 @@ public class InternalResponseCodeService {
 		List<Long > paymentProcessorResponseCodeIds = paymentProcessorInternalResponseCodeDAO.findPaymentProcessorInternalResponseCodeIdsByInternalResponseCode(id);
 		
 		internalResponseCodeDAO.delete(internalResponseCodeToDelete);
-		paymentProcessorInternalResponseCodeDAO.deletePaymentProcessorResponseCodeIds(paymentProcessorResponseCodeIds);
+		if(paymentProcessorResponseCodeIds != null && !paymentProcessorResponseCodeIds.isEmpty()){
+			paymentProcessorInternalResponseCodeDAO.deletePaymentProcessorResponseCodeIds(paymentProcessorResponseCodeIds);
+		}
+		
 	}
 }
